+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the MegaLinter Configuration extension will be documented
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Handle EXTENDS property and highlight inherited variables in the configuration editor
+- Use `x-doc-key` when mentioned in JSON Schema to build links to documentation
+- Fix duplicate label "Config values"
+- Improve responsiveness of the configuration editor on small screens
+
 ## [0.0.1] - 2026-01-12
 
 - Initial release of MegaLinter Configuration extension